pyproject.toml
[project]
name = "mcp-server-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
